refactor: 컴포넌트명의 Messages 를 MessageGroup 으로 변경 및 props type 컴포넌트 내부로 이동
