+++ src/app/chat/[chatId]/_components/MessageBubble.tsx
