+++ src/app/chat/[chatId]/_components/MessageContent.tsx
